Redo the Dict implementation for lower RAM usage.
The dictionary was taking up too much memory (~90MB). To run on an F1
instance in App Engine, the server will need to take up <128MB.

The dictionary was redone so that each node no longer keeps a dense
array of pointers to all possible children (most elements of which were
nil in a large dictionary). Instead, each node holds a pointer to its
sibling in addition to a pointer to its first child. We hold far fewer
nil pointers in memory this way, at the expense of having to traverse
the list to find the correct child.